change the way to fill out the page
from "justifycontent: space-between" to setting
the children flex: 1 1 0 and having a max-width
to each child.
this is done to prevent unneccary space between
the header and the body of the body.
reserch needs to be done on how proparly and
specifically on why did the footer now
become at the bottom with flex 1 1 0 and a
max-width just like it did
with "justify content: space-between